Implementing split pricelist in RP and NP
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C43421

+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Entities/SalePricelistChanges/SalePriceListCustomerChange.cs
@@ -9,7 +9,7 @@ namespace TOne.WhS.BusinessEntity.Entities
     public class SalePriceListCustomerChange
     {
         public long BatchId { get; set; }
-        public int PriceListId { get; set; }
+        public long PriceListId { get; set; }
         public int CountryId { get; set; }
         public int CustomerId { get; set; }
     }
