+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Entities/SalePricelistChanges/SalePriceListSnapshot.cs
@@ -8,7 +8,7 @@ namespace TOne.WhS.BusinessEntity.Entities
 {
     public class SalePriceListSnapShot
     {
-        public int PriceListId { get; set; }
+        public long PriceListId { get; set; }
         public SnapShotDetail SnapShotDetail { get; set; }
     }
     public class SnapShotDetail
