+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Entities/SalePricelistChanges/SalePricelistRPChange.cs
@@ -9,7 +9,7 @@ namespace TOne.WhS.BusinessEntity.Entities
     public class SalePricelistRPChange
     {
         public int CountryId { get; set; }
-        public int PriceListId { get; set; }
+        public long PriceListId { get; set; }
         public string ZoneName { get; set; }
         public long? ZoneId { get; set; }
         public int RoutingProductId { get; set; }
